+++ src/main/java/qastudio/backend/domain/auth/service/AuthCommandService.java
@@ -1,0 +1,9 @@
+package qastudio.backend.domain.auth.service;
+
+import qastudio.backend.domain.auth.dto.request.AuthRequest;
+import qastudio.backend.jwt.TokenInfo;
+
+public interface AuthCommandService {
+    void userSignUp(AuthRequest request);
+    TokenInfo localLogin(AuthRequest loginRequest);
+}
