+++ src/main/java/qastudio/backend/domain/auth/service/AuthCommandServiceImpl.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import qastudio.backend.domain.auth.converter.AuthConverter;
 import qastudio.backend.domain.auth.dto.request.AuthRequest;
-import qastudio.backend.domain.user.entity.AccountTable;
 import qastudio.backend.domain.user.entity.User;
 import qastudio.backend.domain.user.entity.enums.EmailType;
 import qastudio.backend.domain.user.repository.AccountTable.AccountTableRepository;
@@ -26,22 +25,22 @@ import qastudio.backend.jwt.TokenInfo;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class AuthServiceImpl implements AuthService {
+public class AuthCommandServiceImpl implements AuthCommandService {
 
     private final PasswordEncoder passwordEncoder;
-    private final AccountTableRepository accountTableRepository;
     private final UserRepository userRepository;
     private final CustomUserDetailsService customUserDetailsService;
+    private final AuthQueryService authQueryService;
     private final JwtTokenProvider jwtTokenProvider;
-    private final AuthConverter authConverter;
+    private final AuthConverter signUpRequestConverter;
 
     @Override
     public void userSignUp(AuthRequest request) {
-        if (existsEmail(request.getEmail())) {
+        if (authQueryService.existsEmail(request.getEmail())) {
             throw new BadRequestException(ErrorStatus.ALREADY_EXIST_EMAIL);
         }
 
-        User user = authConverter.toUser(request);
+        User user = signUpRequestConverter.toUser(request);
         userRepository.save(user);
     }
 
@@ -63,7 +62,7 @@ public class AuthServiceImpl implements AuthService {
             );
 
             // 사용자 ID 조회
-            Long userId = findUserIdByEmailAndEmailType(loginRequest.getEmail(), EmailType.LOCAL);
+            Long userId = authQueryService.findUserIdByEmailAndEmailType(loginRequest.getEmail(), EmailType.LOCAL);
 
             // 토큰 생성 및 반환
             return jwtTokenProvider.generateToken(userId, authentication, false);
@@ -72,17 +71,5 @@ public class AuthServiceImpl implements AuthService {
         } catch (BadCredentialsException ex) {
             throw new BadRequestException(ErrorStatus.INVALID_PASSWORD);
         }
-    }
-
-    @Transactional(readOnly = true)
-    public boolean existsEmail(String email) {
-        return accountTableRepository.existsByEmail(email);
-    }
-
-    @Transactional(readOnly = true)
-    public Long findUserIdByEmailAndEmailType(String email, EmailType emailType) {
-        AccountTable account = accountTableRepository.findByEmailAndEmailType(email, emailType)
-                .orElseThrow(() -> new BadRequestException(ErrorStatus.USER_NOT_FOUND));
-        return account.getUser().getId();
     }
 }
