+++ src/main/java/qastudio/backend/domain/auth/service/AuthQueryService.java
@@ -1,0 +1,8 @@
+package qastudio.backend.domain.auth.service;
+
+import qastudio.backend.domain.user.entity.enums.EmailType;
+
+public interface AuthQueryService {
+    boolean existsEmail(String email);
+    Long findUserIdByEmailAndEmailType(String email, EmailType emailType);
+}
